refactor: extract focus trap hook from feed button
