+++ apps/web/src/app/__tests__/collections-new.test.tsx
@@ -205,14 +205,18 @@ describe("NewCollectionPage", () => {
       target: { value: "bad--slug" },
     });
 
-    expect(await screen.findByText("※ 3-40文字, 英小文字/数字/ハイフン")).toBeInTheDocument();
+    expect(
+      await screen.findByText("※ 3-40文字, 英小文字/数字/ハイフン"),
+    ).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "作成" })).toBeDisabled();
   });
 
   it("shows API error when create fails", async () => {
     vi.mocked(apiFetch).mockImplementation(async (url, init) => {
       if (url === "/api/users/user-1/collections") {
-        return new Response(JSON.stringify({ collections: [] }), { status: 200 });
+        return new Response(JSON.stringify({ collections: [] }), {
+          status: 200,
+        });
       }
       if (url === "/api/collections" && init?.method === "POST") {
         return new Response(JSON.stringify({ error: "already exists" }), {
@@ -228,7 +232,9 @@ describe("NewCollectionPage", () => {
       target: { value: "Lab Picks" },
     });
 
-    await waitFor(() => expect(screen.getByText("✓ 使用可能")).toBeInTheDocument());
+    await waitFor(() =>
+      expect(screen.getByText("✓ 使用可能")).toBeInTheDocument(),
+    );
     fireEvent.click(screen.getByRole("button", { name: "作成" }));
 
     expect(await screen.findByText("already exists")).toBeInTheDocument();
